AVRO-194. C++ varint encoding buffer too small.
git-svn-id: https://svn.apache.org/repos/asf/hadoop/avro/trunk@835508 13f79535-47bb-0310-9956-ffa450edef68

+++ src/c++/api/Reader.hh
@@ -146,7 +146,7 @@ class Reader : private boost::noncopyable
         int shift = 0;
         do {
             in_.readByte(val);
-            uint64_t newbits = (val & 0x7f) << shift;
+            uint64_t newbits = static_cast<uint64_t>(val & 0x7f) << shift;
             encoded |= newbits;
             shift += 7;
         } while (val & 0x80);
